explicitly convert log() results to quite icc
git-svn-id: https://qthreads.googlecode.com/svn/trunk@1461 9dbd018f-d1da-d342-4c11-58c73dfe904b

+++ src/qutil.c
@@ -323,7 +323,7 @@ static void drf_qsort(void *const array, const size_t elements,
 
 static void drf_qsort_dbl(double *const arr, const size_t elements)
 {
-    const size_t MAX = log((double)elements) + 5;
+    const size_t MAX = (size_t) log((double)elements) + 5;
     ssize_t beg[MAX], end[MAX], i = 0, L, R, swap;
     double piv;
 
@@ -365,7 +365,7 @@ static void drf_qsort_dbl(double *const arr, const size_t elements)
 
 static void drf_qsort_algt(aligned_t * const arr, const size_t elements)
 {
-    const size_t MAX = log((double)elements) + 5;
+    const size_t MAX = (size_t) log((double)elements) + 5;
     ssize_t beg[MAX], end[MAX], i = 0, L, R, swap;
     aligned_t piv;
